+++ sources/ScreenshotAnnotator/Services/ApplicationLinks.cs
@@ -5,5 +5,4 @@ public class ApplicationLinks
     public const string LatestRelease = "https://github.com/drweb86/annotator/releases/latest";
     public const string ApiLatestRelease = "https://api.github.com/repos/drweb86/annotator/releases/latest";
     public const string AboutUrl = "https://github.com/drweb86/annotator";
-    public const string LicenseUrl = "https://github.com/drweb86/annotator/blob/main/LICENSE";
 }